requirements.txt
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.0.0
openai>=1.0.0
PyYAML>=6.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
psutil>=5.9.0
httpx>=0.24.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.28.0
qdrant-client>=1.6.0
sentence-transformers>=2.2.0
faster-whisper>=0.9.0
kokoro-onnx>=0.1.0
docker>=6.1.0
beautifulsoup4>=4.12.0
opencv-python>=4.8.0
pyautogui>=0.9.54
google-api-python-client>=2.90.0
msal>=1.22.0
pillow>=10.0.0
